+++ resources/GUIDE.md
@@ -1,5 +1,5 @@
 Inside of `webpack.config.js`, update the `module.exports` object as such:
 
-//# add_webpack_config:webpack.config.js:3,6,9
+//# refactor_tests:resources/webpack.config.js:3,6,9
 
 You can specify the output of the bundle like this.
