+++ src/parse-article.js
@@ -1,4 +1,7 @@
-const parseSnippet = require("./parse-snippet")
+const {
+  getSnippetDetails, 
+  parseSnippet
+} = require("./parse-snippet")
 const simpleGit = require("simple-git/promise")(__dirname)
 
 function show(branch, file) {
@@ -13,22 +16,21 @@ async function processArticle(article) {
     const line = lines[i]
 
     if (line.substr(0, 3) === "//#") {
-      const {branch, file, lineNumbers} = parseSnippet(line)
+      const {branch, file, lineNumbers} = getSnippetDetails(line)
 
       try {
         const text = await show(branch, file)
         const snippet = parseSnippet(text, lineNumbers)
         output = output + snippet
-        await writeFile(`snippet${i}.md`, snippet)
       } catch (e) { 
-        console.log(e)
+        console.log("Error", e)
       }
-
-
     } else {
       output = output + line + "\n"
     }
   }
+
+  return output
 }
 
 module.exports = processArticle
